Rename files to child metering points
